Refactor API client and dashboard handling
The commit refractors the DashboardProvider component and enhances the governance of server services. It uses API client methods in lieu of the previously used useDashboardServices hook and makes adjustment to the server-side 'updateDashboard' method to make the updating workflow more efficiently with transparency in regards to the process status. Other minor changes made improve performance and readability.